[refactor] auth component re export
