+++ src/components/Auth/index.ts
@@ -1,0 +1,6 @@
+export { default as AuthBottomInfo } from './AuthBottomInfo';
+export { default as AuthFind } from './AuthFind';
+export { default as AuthSelectButton } from './AuthSelectButton';
+export { default as AuthSubmitButton } from './AuthSubmitButton';
+export { default as Input } from './Input';
+export { default as ResetPasswordButton } from './ResetPasswordButton';
