+++ src/page/auth/Member.tsx
@@ -8,15 +8,14 @@ import {
   IAuthLogInByEmailApiResponse,
   IEmailSignInInputData,
 } from 'types/Auth';
-import Input from 'components/Auth/Input';
 import { Link } from 'react-router-dom';
-import AuthSubmitButton from 'components/Auth/AuthSubmitButton';
 import useValidErrors from 'hooks/useValidErrors';
 import { useApi } from 'hooks/useApi';
 import { logInByEmail } from 'api/auth';
 import useApiNavigation from 'hooks/useApiNavigation';
 import { useAppDispatch } from 'redux/store';
 import { setToken } from 'redux/reducer/auth';
+import { AuthSubmitButton, Input } from 'components/Auth';
 
 function Member() {
   const {
