+++ src/page/auth/SignUp.tsx
@@ -5,12 +5,11 @@ import { ReactComponent as ValidationErrorSvg } from '@svg/validation-error.svg'
 import { useForm } from 'react-hook-form';
 import { IEmailSignUpInputData, IJoinByEmailApiResponse } from 'types/Auth';
 import { emailRegex } from '../../utils/index';
-import Input from 'components/Auth/Input';
-import AuthSubmitButton from 'components/Auth/AuthSubmitButton';
 import useValidErrors from 'hooks/useValidErrors';
 import { useApi } from 'hooks/useApi';
 import { joinByEmail } from 'api/auth';
 import useApiNavigation from 'hooks/useApiNavigation';
+import { AuthSubmitButton, Input } from 'components/Auth';
 
 function SignUp() {
   const {
